refactor: move FiLM module under models/layers and remove re-export
