--- graph_weather/film.py
@@ -1,3 +1,0 @@
-from .models.film import FiLMApplier, FiLMGenerator
-
-__all__ = ["FiLMGenerator", "FiLMApplier"]
